Lock dependencies to pulumi branches
Locks our dependencies on the TF AWS provider to a pulumi
fork and our dependency on pulumi/lumi to the tfbridge branch.

Updates the AWS provider to align with current pulumi-master
branch on pulumi/terrarofm-provider-aws.

Also moves .gitignore to the root so that we can safely blow away
and regenerate the entire pack/ folder when needed.