Add webserver integration test variant
This adopts the new integration test changes, and uses them to perform
an update that entails a replacement as part of the AWS webserver test.
All it does is changes the security group's description, plus adds an
ingress rule for port 22, triggering a replacement deployment.

In the future, we probably want to add lots more variants to trigger
interesting code paths (like updates that don't trigger replacements).