Use dependsOn to manually link API Gateway resources
Before creating an `apigateway.Integration`, we must have created the
`apigateway.Method`s associated with the target `apigateway.RestApi`.
Similarly, we must have created all of those things before creating
the final `apigateway.Deployment` resource.  Because the properties of
said resources do not naturally lead to the proper dependency graph,
however, enabling parallelism can lead to attempting to create the
resources out of order.  To fix this, we manually link them with
dependsOn.  Please see pulumi/pulumi-aws#22 for ideas on improving this.