Update to latest AWS terraform provider (#37)
We have adopted 1.0.0 of the AWS terraform provider, but also pulled in a PR adding support for NLB, which appears to be targeting a 1.0.1 release. (see terraform-providers/terraform-provider-aws#1629)

Naming is a little strange on these load balancers due to confusing  decisions by both AWS and Terraform.  For our modules mapped on top - we use:

* `elasticloadbalancing` module supports ELB
* `elasticloadbalancingv2` module supports ALB and NLB
* `applicationloadbalancing` module supports ALB

CloudFormation uses the first two names.  Terraform uses an `elb` prefix for the first, an `lb` prefix for the second and an `alb` prefix for the third.  They have effectively deprecated the third as part of the addition of the second group (in fact, those are just aliases now I believe).