Permit bucket names, alongside strongly typed Buckets (#104)
In cross-stack linking, often you just have a name from a configuration
variable.  Not having a Bucket in hand meant there was no way to populate
such a bucket with objects.  This change adopts a new feature in the TF
bridge where we can use union types to accept both.