Consume NPM packages in our examples (#125)
This change switches our examples over to using real NPM packages.
It uses the `"latest"` tag so that we always download the latest
package.  There is currently a known issue which is that we won't
actually land on the latest due to the Yarn lock files, however,
we will be addressing that shortly with updates to our test framework.