Publish with real @pulumi/pulumi dependencies (#139)
This change adopts our new workflow for publishing a @pulumi/aws
package to NPM that can be used entirely with npm install, and notably
no symlinking.  Because all of our tests still use symlinking *within*
the SDK, we still use peerDependencies in most places.  Upon publishing,
we will promote the peerDependency of @pulumi/pulumi to a real one.