Support configuring the paths to include in serverless.Functions (#210)
An `aws.serverless.Function` can now accept a list of include paths, which will be included inside the Lambda package along with the generated `__index.js` file

Also changes the default from '.' to './node_modules', so that the very common case of changes to the root folder do not trigger changes to the code for lambdas.  If files from the local folder are needed by the lambda, they can be included in the `includePaths` parameter.

Also adds support for providing a pre-created Role instead of just Policies, to support scenarios where users want to pre-create or share a Role across multiple functions.

Fixes #35.