Ensure install script is added before publishing
In #226 we removed calling promote.js when publishing, as we no longer
used peerDependencies for `@pulumi/pulumi`. However, that script was
also responsible for adding an install script to the package.json so
when installed, we would call `pulumi plugin install`. We need to
retain this behavior until we stop using tfgen to generate the
package.json and can add it by hand (or if we decide we want to
continue to use tfgen to write the package.json file, fix tfgen to do
add the install script).

Fixes #229