Ensure worktree is clean after build
Because we now check in sources generated by tfgen, it's possible that
the sources could get out of sync, since tfgen regenerates this every
run. This is common if the reference to the provider is updated in
Gopkg.toml and a build is not done afterwards.

In this case, we'll do two bad things:

1. We'll publish the plugin with a -dirty tag
2. The actual source in the tree will not reflect reality.

So, in Travis, let's check that the tree is clean after building on
push and as the very last step of a PR.