Use version from package.json for the plugin
Explicitly version the plugin based on what is present in the
package.json file for nodejs package. This is the canonical version we
use and we shouldn't generate it again. This is a belt and suspenders
thing but ensures the version we publish exactly matches what we
embedded into the packages for a `pulumi install`.