Detect org-scoped packages for functions
The logic in place to detect org-scoped packages needed to work around
the fact that names for them weren't correct, and did so by parsing the
path name.  Unfortunately, the regex including a leading slash, which
meant that the subsequent name match would fail.  I also *think* the
logic would have failed on Windows, but am not entirely sure about this.

Instead of fixing the regex, I've added logic to split the path into
parts and then match the last two elements of the path.  I believe this
accomplishes the intent while being a bit more resilient to pathing
issues and hopefully working on Windows.

I'd like to add tests for this but we don't have a great story for how
to unit test overlays like this at the moment, so will look into doing
that as a follow up.