Explicilty map tag => tags on aws.autoscaling.Group
We have an awkward mapping of tag=>tags and tags=>tagsCollection on `aws.autoscaling.Group`.  This leads to some confusion in our tfbridge mapping.

For now, we can fix #293 by making the tag=>tags part of the mapping explicit.  Longer term, we likely need to make deeper improvements in tfbridge to handle this and/or come up with a different approach for aws.autoscaling.Group projection of tag/tags.

Fixes #293.