Work around test hangs in Travis
The combination of a high level of parallelism and having a lot of
tests to run at the same time has lead to cases in Travis where we
don't see output from `go test` for a while, causing the job to be
killed.

Work around this in two ways, first, bump down the level of
parallelism (the travis workers have two cores, so running 10 tests in
parallel is kind of a stresser for it) and split the test of tests
across two invocations of go test.

This seems to ameliorate the problem.