startingPositionTimeout => string in Kinesis mixin
During upgrade of the Terraform provider, an attribute got added to the
aws.lambda.EventSourceMapping which defines startingPositionTimeout as a
string in RFC3339 format rather than a number as it was previously
defined in this mixin.

This commit changes the mixin to use a string and updates the comments
comment to reflect that. This fixes the build error when compiling
TypeScript after running tfgen with AWS Provider version 1.49.0.

Note that this is a breaking change to the public API and will want
noting in CHANGELOG.