Don't use aws.Tags for aws.autoscaling.Group::tagsCollection
The existence of tagsCollection on aws.autoscaling.Group is to work around a
limitation of HCLv1 configuration for Terraform, and while it is not the
preferable way to define tags in Pulumi, is sometimes generated by tf2pulumi.

This commit fixes the type generation such that a list of maps with keys
consisting of `key`, `value` and `propagate_at_launch` will work for tagging an
autoscaling group.

Use of `tags` continues to be preferred.