Remove aws.Tags overlay from resources
Using `{ [key: string]: string }` as the type for `aws.Tags`, and projecting
that for the `tags` property on AWS resources precludes tag values from being
Outputs, despite the tags that the engine supports that. Until a better solution
exists, this commit removes the overlay types from the various resources.

The tags overlay and the `aws.Tags` type is kept, in case people have references
to it as (for example) a variable type - it should still be compatible with the
new definition.

Fixes #410.