Add dimension merging. (#521)
This is valuable for when we have instance methods on resources that expose metrics. These will have set some dimension of the metric (for example, the 'FunctionName' for an aws lambda metric). However, if the user wants to go set other dimensions, we don't want to blow away the existing dimension.