Strongly-typed S3 bucket website routingRules
This augments the `routingRules` property to accept a strongly-typed
array of objects, which helps to know which properties can be set and
avoids the need to use `JSON.stringify`. This is especially helpful
as some properties can only be set if other properties are not set.

Depends on a change in `pulumi-terraform` to allow JSON arrays in
`TransformJSONDocument`.

Details:
https://www.terraform.io/docs/providers/aws/r/s3_bucket.html#routing_rules
https://docs.aws.amazon.com/AmazonS3/latest/dev/how-to-page-redirect.html#advanced-conditional-redirects (scroll down for schema)