Update `ContainerDefinition` to latest ECS definition
Align with https://docs.aws.amazon.com/AmazonECS/latest/APIReference/API_ContainerDefinition.html.

The change in type for `dependsOn` is a TypeScript-level breaking change, but the previous type was incorrect and would not have worked with ECS, so it is not in practice a breaking change for any working user code.