Avoid baking config from environment into checkpoint (#894)
In https://github.com/pulumi/pulumi-aws/pull/874 we added config defaults from environment variables for four new configuration settings.  Thees config defaults are used in two places: (1) `aws.config` and (2) the defaults for `aws.Provider`.  For (1) these changes were a good thing, but for (2) they led to values from the environment getting baked into checkpoints that should not be.

It's not clear to me that we should be doing (2) at all - that is - I don't think `region` or `profile` should be picked up from the environment and baked into the checkpoint file either.  But for now we'll just revert the more recent change here which has led to the more significant immediate issue.

Part of https://github.com/pulumi/pulumi-aws/issues/814.

Fixes #890.