utils.ts causes a circular reference
An example of this is cloudwatch eventRuleMixin, which references lambda.ts, which imports utils.ts, which imports the root mixin file, which requires everything.

This causes a runtime error with classes.

I am unsure about the impacts of removing this, as far as I can tell it is imported from index.ts, meaning if you import `@pulumi/aws` the side effect of adding the proxied .sdk will still happen. But there may be other things I cannot spot