Change how preConfigureCallback happens to ensure Metadata API included
Fixes: #1995

We now ensure that the preConfigureCallback function is sensitive to
skipCredentialsValidation and skipMetadataApiCheck

If the user sets skipCredentialsValidation: true in their config
then we don't run any preConfigureCallback. It's an explicit operation
to skip the credentials check.

If the user is validating via the metadata api, then they will set
skipMetadataApiCheck: false in the configuration and we need to ensure
we set the correct imds confiuration before we try and authenticate
the user