Aws `@types/aws-lambda` for lambda callback signatures.
We currently hard code our own TypeScript types for the callback functions.  But these are well-defined in a standard way in the ecosystem in `@types/aws-lambda`.  We can instead use that library as a type-only dependency.

This fixes one bug, and ensures greater compatibility with TypeScript libraries that define Lambda callbacks.

Fixes #1314.