Adds a phone make target to check local schema against master
The idea behind this addition is to avoid the CI round-trip for
ensuring our currently generated schema produces the results that we expect.
Currently missing from schema-tools is an implementation of the version command
which means we'll download a new artifact every time we run.

We also don't have a very safe OS check, etc, but it will at least work
on apple machines for the time being.