Cache Go builds on sdk/go.sum (#2483)
This is an unusual repository - go.sum changes all the time, while there are not so many parallel builds. This means
that typically all PRs have cache misses because they edit some go.sum somewhere. It would be preferable to fetch the
wrong build cache, hoping some parts of it still speed up the build, than it is to accept a cache miss.

This PR relies on the fact that sdk/go.sum is fairly stable, and uses it as a key. This is less precise but might be
preferable.