Remove aws.sdk property
The `aws.sdk` proeprty was introduced several years ago to offer a convenient way for callbacks using function serialization to get ahold of the AWS SDK which is available within the Lambda runtime.

However, the AWS SDK v2 that was exposed here is being deprecated, and has been emitted a deprecation warning for some time for anyone using this via Pulumi.

The AWS SDK v3 doesn't expose a monolitchic module we can expose here - providing instead many seperate SDKs for each service.

So we expect the best path is to drop this from the AWS provider in the upcoming major version release.

Users who were using it in <6.0.0 versions will need to move to using `@aws-sdk/client-s3` or similar libraries directly.