Fix defaultTags to trigger downstream updates
The `defaultTags` input property on providers now triggers an update on
downstream resources. Because of the nature of the fix in the Pulumi TF Bridge,
and to limit the scope of the change, we mark only the `tags_all` field from TF
schema as a `ComputedInput`, triggering new diff behavior within the bridge.

As the `tags_all` field is otherwise hidden as a "computed" value, we omit
detailed diff information. This is a tradeoff to avoid a more invasive change
with more risk.

Fixes #1655.

Example output:

```
pulumi up --skip-preview
Updating (test-aws-1655)

View in Browser (Ctrl+O): https://app.pulumi.com/pulumi/simple-yaml/test-aws-1655/updates/9

     Type                     Name                       Status              Info
     pulumi:pulumi:Stack      simple-yaml-test-aws-1655
 ~   ├─ pulumi:providers:aws  aws-provider               updated (0.13s)     [diff: ~defaultTags]
 ~   ├─ aws:s3:Bucket         my-bucket                  updated (0.87s)
 ~   └─ aws:s3:BucketObject   index.html                 updated (0.24s)
```

Verified when adding, updating, deleting, and removing entirely the
`defaultTags` property.