Compute `"tags"` in Check (#2655)
* Add GRPC test to replicate 2633

* Test that we set "tags" on creation

TF does, and if we assume that "tags" is set correctly, we then get the expected diff.

* Ensure valid diffs by overriding check

* Remove `.XComputeInput`

This fix takes effect during Check, which occurs before Diff. We don't need to do anything
special for Diff now.

* Depend on https://github.com/pulumi/pulumi-terraform-bridge/pull/1310

* Turn tests back on

* Consume upstream changes

The previous fix works on all resources... except our s3legacy bucket because it was not
updated to use upstream's new tags strategy. I have changed our fork to use the new
strategy.

The change is here:
https://github.com/pulumi/terraform-provider-aws/compare/patched-v5.9.0...patched-v5.9.0-with-modern-s3legacy-tags?expand=1

* Explain .Mappable() has the same shape as NewTagConfig expected

* Template out YAML Tests

* Accommodate aws:cognito:UserPool

* Switch to environment for pf

* Add a test for aws:s3:BucketV2

We are seeing tests fail for aws:cognito:UserPool but I know this works for
aws:s3:BucketV2. These are both SDKv2 resources. Since they work differently I have
ensured that both are tested.

* Skip aws:cognito:UserPool test

This resource is broken upstream, so we skip the test here.

* Remove outdated tests

I have left skeleton code for GRPC based testing, as I often find it useful when debugging
locally. I have noted that it does not run in CI, and is just there for local development.

* Don't pass empty tags properties