Fix panics in applyTags when tags are unknown (#2776)
Fixes https://github.com/pulumi/pulumi-aws/issues/2774

We detected an issue in pulumi-eks test suite against the latest v6.0.3 release. This theoretically could affects users that set resource and or provider tags to unknown Outputs, although I tried and failed to reproduce outside of the pulumi-eks test.

The fix for https://github.com/pulumi/pulumi-aws/issues/1655 included reimplementing tag merging (computing effective tags from the set of provider-level tags and resource-level tags) on Pulumi side as a callback to the Check Pulumi gRPC method. This works against resource.PropertyValue model and did not fully account for receiving a Computed when tags are unknown, resulting in a panic in the eks test.

The fix returns unknowns if any of the tag inputs are unknown. It also puts asserts in for absence of secrets, as the provider does not expect secret values to be passed to Check and instead handled by Pulumi CLI.