Add a test
This test is forward looking: right now, the bridge panics if we try to log without a
valid resource host. This fails the test, which means the test will fail if we try to log
a deprecation message. When we prevent the bridge from panicking on an empty host, the
test will still fail if we regress.