Disable slow Plugin Framework Config validation (#2920)
Fixes #2880 

Our Plugin Framework integration has degenerate performance when
validating custom endpoints, affecting the users of the AWS Provider.
This is now tracked in
https://github.com/pulumi/pulumi-terraform-bridge/issues/1448 but not
yet fully resolved. In the meantime, this change disables provider-level
config validation with Plugin Framework. For this provider this is safe
to do because SDKv2-style validators are still checking the
user-provided data and generating error messages. Unit tests are added
to ensure that remains to be the case. A small runtime time-check is
added to make sure performance of CheckConfig with custom endpoints does
not regress again.