adds support for aws arm64 batch jobs in v5
This updates the AWS Go SDK to allow the container properties section in
the `aws.batch.JobDefinition` to support the ARM64 architecture.

This property is untyped, so it doesn't exist in the schema. Prior to
this version of the AWS Go SDK, the property would silently drop
unsupported fields, so this allows users to use these properties without
having to update to v6