Update AWS Go SDK (#3031)
Adapts https://github.com/pulumi/pulumi-aws/pull/3021 to make it
releasable. Tests added as Python acceptance tests. CI configs tweaked
to run the tests, they are now passing.

Original description below.

This updates the AWS Go SDK to allow the container properties section in
the aws.batch.JobDefinition to support the ARM64 architecture.

This property is untyped, so it doesn't exist in the schema. Prior to
this version of the AWS Go SDK, the property would silently drop
unsupported fields, so this allows users to use these properties without
having to update to v6