Fixes inability to tag AWS Firehose stream (#3065)
Fixes #3064 

There is an issue with users being unable to add tags to AWS Firehose
delivery stream and select other resources.

As part of the https://github.com/pulumi/pulumi-aws/pull/2944 mitigation
the provider needs to make sure that tags_all attributes, if present,
are not Computed because pulumi-aws will set them before passing down to
TF code.

TestTagsAllNoLongerComputed was supposed to check this.

Unfortunately the transformation missed that some resources define
SchemaFunc instead of Schema, and so did the test. The test is made
stricter here, and the underlying transformation is fixed.