Optimize startup performance (#3070)
Zooming in a lot on AWS Provider startup costs, noticing a recurring
pattern. In SDKv2 we have schema.Resource that has Schema of type map
but also SchemaFunc The latter is designed specifically to defer / lazy
instantiate schemas that are big. It looks like in AWS QuickSight and
some other very large schemas make use of this. However over time
transformations baked into upstream providers and our own bridge added
up to a lot of passes that force-call these SchemaFunc() on startup,
even if you don't use QuickSight. When done sequentially it all adds up.
Moreover since there's no true laziness we end up paying the QuickSight
etc cost multiple times.

This change does the following to avoid `SchemaFunc()` calls:

- turn off upstream tags check - this seems to be a validation-only
measure, that somehow manages to call SchemaMap and SchemaFunc and slow
things down; it now disabled in the bridged provider

- rewrite markTagsAllNotComputedForResources to avoid calling
`SchemaMap()` eagerly; instead call on-demand and try to cache the call
so QuickSight and friends are not initialized twice