Stop embedding bridge-metadata.json
At this point bridge-metadata.json is a compile-time only concern as it is not needed at runtime at
all. This PR makes the change obvious and stops embedding it. This should help the binary size a
bit, possibly to the tune of 30MB (uncompressed):

 31M    provider/cmd/pulumi-resource-aws/bridge-metadata.json
 88K    provider/cmd/pulumi-resource-aws/runtime-bridge-metadata.json

Only runtime-bridge-metadata.json is currently needed.