Unpin pulumi cli 3 (#3191)
Before this change, we used an outdated Pulumi CLI version to test the
provider against. With this change we start using the latest 3.x release
like most other bridged providers. While it would be my preference to
have it pinned and automatically updating via PRs to move toward
[Hermetic Builds](https://github.com/pulumi/ci-mgmt/issues/738) that's
more work and for now this change is a quick improvement.