Make integration tests stricter (#3183)
Prior to this change, all example-based acceptance tests inherited a
relaxed configuration.

```go
	return integration.ProgramTestOptions{
		ExpectRefreshChanges: true,
		SkipRefresh:          true,
		Quick:                true,
	}
```

This was masking issues with unexpected diffs during `refresh` and any
possible regressions as there provider evolves. After the change, only
the test that need it are configured with these relaxed options and the
code is cross-correlated to root cause issues where appropriate.