Update aws.lambda.Runtime to latest AWS support (#3261)
Catches our enum up to date with the latest from
https://docs.aws.amazon.com/lambda/latest/dg/lambda-runtimes.html.

Also updates the deprecated runtimes to latest, including some which
have been very recently deprecated (but in a few cases, are not yet
erroring on create/update - though those are coming in the next few
weeks).

Also fixes CallbackFunction in TS to ensure it can accept a normal
`string` as well, to ensure that users are not prevented from passing
not-yet-supported enum values in the future. This aligns with the
underlying `Function` type which accepts both.

Fixes #3260.
Fixes https://github.com/pulumi/pulumi-aws/issues/2994.