Remove VPC from tags test matrix (#3296)
Nothing that the TestRandomTagsCombinationsGo test periodically flakes
up, as in https://github.com/pulumi/pulumi-aws/pull/3292 on testing VPC
resources. I've not established the root cause but it appears to be that
the VPCs allocated by independent test instances collide. This is
probably just not a good resource to test tag correctness on and I think
we're not losing much coverage by simply removing it. Buckets tests
still test the SDKv2 based resources.