Revert upstream changes triggering LB panic (#3426)
This reverts the following upstream PRs:
https://github.com/hashicorp/terraform-provider-aws/pull/35671
https://github.com/hashicorp/terraform-provider-aws/pull/35678
as a quick fix to mitigate
https://github.com/pulumi/pulumi-aws/issues/3421 until we root-cause it.

Details on my findings so far:
https://github.com/pulumi/pulumi-aws/issues/3421#issuecomment-1943569655
It looks to me like the issue originates somewhere in our handling of
nulls/empty in the bridge, so seems unlikely to get fixed today.

It also adds a test for LB listeners. I've verified that the test
triggers the panic without the patches and that the patches resolve it.