Patch lambda to allow imports (#3420)
Should address https://github.com/pulumi/pulumi-aws/issues/2392

We regressed in our handling of imported lambdas in 5.28, when we picked
up upstream
https://github.com/hashicorp/terraform-provider-aws/releases/tag/v4.51.0
with
https://github.com/hashicorp/terraform-provider-aws/pull/28963/files,
which introduces the `ExactlyOneOf` constraints in the lambda resource.

This PR reverts the upstream `ExactlyOneOf` constraints and replaces
them with the previously applied `ConflictsWith`. This in turn allows
imports of lambdas via `pulumi import` and the `import` resource option
to work properly.

I've added a GRPC test for a lambda imported via `pulumi import` as well
as an integration test for a lambda imported via the `import` resource
option and a test which checks that we still fail to create lambdas
without any code-related properties (which should be the only case that
now passes the validation step which previously didn't).

Note that this slightly worsens the behaviour in the case when none of
the code-related properties are specified. Previously that'd tirgger a
failure during `Check` and print a sensible error message but now we
will fail during `Create` with the following much less legible error:

```
  aws:lambda:Function (testLambda):
    error: 1 error occurred:
        * creating Lambda Function (testLambda-e9e5e22): operation error Lambda: CreateFunction, https response error StatusCode: 400, RequestID: 88b5f9d7-42d6-4a3b-becc-96cdf954f596, api error ValidationException: 2 validation errors detected: Value '' at 'code.s3Key' failed to satisfy constraint: Member must have length greater than or equal to 1; Value '' at 'code.s3Bucket' failed to satisfy constraint: Member must have length greater than or equal to 3
```

I don't see a sensible way around it though and I think the change is
still worthwhile.


Summary of effects:
| | <5.28 | 5.28-6.22 | This PR |
| :-------------  | :-------: | :----:| :-------:|
| pulumi import  | works? | doesn't work | works|
| import option   | works |doesn't work | works|
| no code lambda| errors during update?| errors during validation |
errors during update |
| other lambdas| works| works  | works |