fix: role inline policies fail to create without a policy name
In order to create an `inline_policy` on a role, both the `policy` and
the `name` fields are required. If only one is provided then the policy
will not be created (and no error will be thrown). Both properties must
be optional though to allow providing empty values, i.e.

```ts
inlinePolicies: []
// or
inlinePolicies: [{}]
```

This PR sets the `name` property to be auto named and adds a transform
on `inline_policy` to only create policies if a `policy` is provided.

fixes #1949