Fix #2442 (#3638)
Fixes #2442 

This adds a diff customizer that ignores changes to parameters that only
change the apply_method and not the value for the aws_db_parameter_group
resource. To make this work, the change also needs to modify the set
element hashing function to identify parameters that differ only by
apply_method as identical.

As a side-effect of the set hashing change, upgrading stacks to the
newer version of the provider with this change will show a reordering
update diff on the ParameterGroup resource.