run unit tests in separate job (#3666)
This pull request adds the `provider/` unit tests as a separate job via
`.ci-mgmt.yaml` ExtraTests config.

It only runs python and nodejs matrix, as there are currently no other
unit tests enabled that need other languages. I'm happy to adjust this
as necessary.

This refactor is meant to address the AWS credentials timeout that
causes #3637.