Fix flaky test TestAccKmsAliasTs
The test is creating a KMS key alias with a static name. When multiple
instances of this test are running in the same AWS account and region they'll race
and occasionally fail because the alias already exists.

This changes the key alias to use a name prefix to guarantee uniqueness.