Fix 0054 patch to pass schema checks
In 0054-Revert-r-aws_db_proxy-Change-auth-from-TypeList-to-T patch we reverted an upstream change to avoid triggering a
panic in Pulumi. The change moved rds.Proxy auths parameter from a List to a Set type. Unfortunately it looks like we
neglected to remove the Set cutomizer that is not a valid customizer for list-typed properties. The resulting schema no
longer passes InternalValidate which the new bridge release will run at make tfgen time, thus blocking the bridge
upgrade.

This should unblock the bridge update.