Fixes errors and resource cycling issues for aws.batch.JobDefinition
Recent changes in hashicorp/terraform-provider-aws#37111 introduced a Diff customizer that leads to errors in the Pulumi
version of the provider due to discrepancies in the order in which Diff customizer functions are applied between Pulumi
and Terraform. This change fixes the problem by applying the experimental PlanResourceChange flag to the affected
resource.

A regression test is included.

See also: https://github.com/pulumi/pulumi-terraform-bridge/issues/1785