Fix empty retry_strategy of Batch JobDefinition causing panics (#3921)
Computing whether a Batch JobDefinition has a diff panics if the retry_strategy is empty. This is a valid configuration because all properties of retry_strategy are optional.

This patch fixes that by adding the missing nil checks for the retry strategy.

What's notable is that having an empty retry_strategy will cause a perma diff. MaxItems=1 blocks with all optional attributes lead to permanent diffs if the user provides an empty block.

Fixes #3905