Fix deletion of inline role policies
In https://github.com/pulumi/pulumi-aws/pull/3587 we introduced
auto-naming for the inline role policies and added filtering of
the inputs within a transformation. This didn't take the special
case of removing all inline policies into consideration, which
requires an empty object as the delete marker.
This change fixes that.