docs: correctly handle `plantimestamp()` function in docs
There is a docs example that is being dropped due to it including a
`plantimestamp()` function which cannot be converted.

This PR puts the function within a string so that it does not attempt to
convert it. I'm open to alternatives if anyone thinks there is something
better to replace it with.

closes #2266