fix!: unable to create fsx for OpenZfs Multi-AZ file system
This PR removes the `MaxItemsOne` config for the `subnetIds` property.
Originally this was a `MaxItemsOne` property in upstream, but it was
removed when Multi-AZ support was added by AWS. In Terraform removing
`MaxItemsOne` is not a breaking change, but for Pulumi it is.

Because it looks like there is very low usage of this resource, and the
change enables a (maybe) more common use case (Multi-AZ filesystem) we
have decided to take the breaking change now.

In order to upgrade users will only need to update their code so that
`subnetIds` is now a list. For example,

From
```ts
const test = new aws.fsx.OpenZfsFileSystem("test", {
    storageCapacity: 64,
    subnetIds: test1.id,
    deploymentType: "SINGLE_AZ_1",
    throughputCapacity: 64,
});
```

To
```ts
const test = new aws.fsx.OpenZfsFileSystem("test", {
    storageCapacity: 64,
    subnetIds: [test1.id],
    deploymentType: "SINGLE_AZ_1",
    throughputCapacity: 64,
});
```

Because we are including the `TransformFromState` function, users should
not need to make any changes to the state themselves.

BREAKING CHANGE: `fsx.OpenZfsFileSystem.subnetIds` now accepts a list
instead of a string

closes #3106, closes #3034